+++ openpifpaf/__init__.py
@@ -18,4 +18,4 @@ from .datasets import DATAMODULES
 from .decoder import DECODERS
 from .network.factory import BASE_FACTORIES, BASE_TYPES, HEAD_FACTORIES, HEAD_TYPES
 from .network.nets import MODEL_MIGRATION
-from .show.painters import PAINTERS
+from .show.annotation_painter import PAINTERS
