+++ openpifpaf/show/__init__.py
@@ -4,4 +4,5 @@ from .animation_frame import AnimationFrame
 from .canvas import Canvas, canvas, image_canvas, white_screen
 from .cli import cli, configure
 from .fields import arrows, boxes, boxes_wh, circles, margins, quiver
-from .painters import AnnotationPainter, CrowdPainter, DetectionPainter, KeypointPainter
+from .painters import CrowdPainter, DetectionPainter, KeypointPainter
+from .annotation_painter import AnnotationPainter
