+++ openpifpaf/visualizer/caf.py
@@ -24,8 +24,7 @@ class Caf(Base):
         super().__init__(meta.name)
         self.meta = meta
         keypoint_painter = show.KeypointPainter(monocolor_connections=True)
-        self.annotation_painter = show.AnnotationPainter(
-                                    painters={'Annotation': keypoint_painter})
+        self.annotation_painter = show.AnnotationPainter(painters={'Annotation': keypoint_painter})
 
     def targets(self, field, *, annotation_dicts):
         assert self.meta.keypoints is not None
